+++ gamac/algorithms/affinity.py
