+++ gamac/algorithms/agglomerative.py
