--- gamac/algorithms/birch_gpu.py
@@ -1,68 +1,0 @@
-import numpy as np
-from numba import cuda
-
-
-class ClusteringFeature:
-    def __init__(self, N=0, LS=None, SS=None):
-        self.N = N  # количество точек
-        self.LS = LS if LS is not None else np.zeros(2)  # линейная сумма
-        self.SS = SS if SS is not None else 0  # сумма квадратов
-
-    def update(self, point):
-        """Обновляет кластерную функцию с новой точкой."""
-        self.N += 1
-        self.LS += point
-        self.SS += np.dot(point, point)
-
-    def merge(self, other):
-        """Объединяет две кластерные функции."""
-        self.N += other.N
-        self.LS += other.LS
-        self.SS += other.SS
-
-    def radius(self):
-        """Вычисляет радиус кластера."""
-        return np.sqrt(self.SS / self.N - np.dot(self.LS / self.N, self.LS / self.N))
-
-
-@cuda.jit
-def update_cf_kernel(points, cf_list, threshold):
-    idx = cuda.grid(1)
-    if idx < points.shape[0]:
-        point = points[idx]
-        # Логика обновления CF должна быть реализована здесь
-        # Для простоты мы просто обновляем первый CF в списке
-        cf_list[0].update(point)
-
-
-class BIRCH:
-    def __init__(self, branching_factor=50, threshold=1.0):
-        self.branching_factor = branching_factor
-        self.threshold = threshold
-        self.cf_list_device = None
-
-    def fit(self, X):
-        """Обучает модель на данных X."""
-        n_samples = X.shape[0]
-
-        # Переносим данные на GPU
-        points_device = cuda.to_device(X)
-
-        # Инициализация списка кластерных функций на GPU
-        cf_list_device = [ClusteringFeature() for _ in range(self.branching_factor)]
-
-        # Переносим кластерные функции на GPU (это упрощение)
-        # В реальной реализации нужно использовать более сложную структуру данных
-
-        self.cf_list_device = cuda.to_device(cf_list_device)
-
-        # Запуск CUDA ядра для обновления CF
-        threads_per_block = 32
-        blocks_per_grid = (n_samples + (threads_per_block - 1)) // threads_per_block
-        update_cf_kernel[blocks_per_grid, threads_per_block](
-            points_device, self.cf_list_device, self.threshold
-        )
-
-    def get_clusters(self):
-        """Получает кластеры из дерева."""
-        return [cf for cf in self.cf_list_device.copy_to_host()]
